+++ backend/src/model/entity/plantings.rs
@@ -20,16 +20,12 @@ pub struct Planting {
     pub x: i32,
     /// The y coordinate of the position on the map.
     pub y: i32,
-    /// The width of the plant on the map.
-    pub width: i32,
-    /// The height of the plant on the map.
-    pub height: i32,
+    /// The size of the planting on the map in x direction.
+    pub size_x: i32,
+    /// The size of the planting on the map in y direction.
+    pub size_y: i32,
     /// The rotation in degrees (0-360) of the plant on the map.
     pub rotation: f32,
-    /// The x scale of the plant on the map.
-    pub scale_x: f32,
-    /// The y scale of the plant on the map.
-    pub scale_y: f32,
     /// The date the planting was added to the map.
     /// If None, the planting always existed.
     pub add_date: Option<NaiveDate>,
@@ -63,16 +59,12 @@ pub struct UpdatePlanting {
     pub x: Option<i32>,
     /// The y coordinate of the position on the map.
     pub y: Option<i32>,
-    /// The width of the plant on the map.
-    pub width: Option<i32>,
-    /// The height of the plant on the map.
-    pub height: Option<i32>,
+    /// The size of the planting on the map in x direction.
+    pub size_x: Option<i32>,
+    /// The size of the planting on the map in y direction.
+    pub size_y: Option<i32>,
     /// The rotation of the plant on the map.
     pub rotation: Option<f32>,
-    /// The x scale of the plant on the map.
-    pub scale_x: Option<f32>,
-    /// The y scale of the plant on the map.
-    pub scale_y: Option<f32>,
     /// The date the planting was added to the map.
     pub add_date: Option<Option<NaiveDate>>,
     /// The date the planting was removed from the map.