+++ frontend/src/features/map_planning/layers/plant/components/PlantingAttributeEditForm.tsx
@@ -1,56 +1,58 @@
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useForm } from 'react-hook-form';
+import { FormProvider, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { z } from 'zod';
 import { PlantingDto } from '@/api_types/definitions';
 import SimpleButton, { ButtonVariant } from '@/components/Button/SimpleButton';
-import SimpleFormInput from '@/components/Form/SimpleFormInput';
+import { DebouncedSimpleFormInput } from '@/components/Form/DebouncedSimpleFormInput';
 import { useFindPlantById } from '@/features/map_planning/layers/plant/hooks/plantHookApi';
-import { useDebouncedSubmit } from '@/hooks/useDebouncedSubmit';
-import CheckIcon from '@/svg/icons/check.svg?react';
-import CircleDottedIcon from '@/svg/icons/circle-dotted.svg?react';
 import { PlantNameFromAdditionalNameAndPlant, PlantNameFromPlant } from '@/utils/plant-naming';
+import { calculatePlantCount, getPlantWidth } from '../util';
 
 const PlantingAttributeEditFormSchema = z
   // The 'empty' value for the API is undefined, so we need to transform the empty string to undefined
   .object({
     addDate: z.nullable(z.string()).transform((value) => value || undefined),
     removeDate: z.nullable(z.string()).transform((value) => value || undefined),
+    sizeX: z.number().int().nullable(),
+    sizeY: z.number().int().nullable(),
   })
   .refine((schema) => !schema.removeDate || !schema.addDate || schema.addDate < schema.removeDate, {
     path: ['dateRelation'],
   });
 
-export type PlantingDateAttribute = Pick<PlantingDto, 'addDate' | 'removeDate'>;
+export type PlantingFormData = Pick<PlantingDto, 'addDate' | 'removeDate' | 'sizeX' | 'sizeY'>;
 
 export type EditPlantingAttributesProps = {
-  onAddDateChange: (addDate: PlantingDateAttribute) => void;
-  onRemoveDateChange: (removeDate: PlantingDateAttribute) => void;
+  onAddDateChange: (formData: PlantingFormData) => void;
+  onRemoveDateChange: (formData: PlantingFormData) => void;
+  onWidthChange: (formData: PlantingFormData) => void;
+  onHeightChange: (formData: PlantingFormData) => void;
   onDeleteClick: () => void;
   isReadOnlyMode: boolean;
 };
 
-export type EditSinglePlantingProps = EditPlantingAttributesProps & {
+export type SinglePlantingAttributeFormProps = EditPlantingAttributesProps & {
   planting: PlantingDto;
 };
 
-export type EditMultiplePlantingsProps = EditPlantingAttributesProps & {
+export type MultiplePlantingsAttributeFormProps = EditPlantingAttributesProps & {
   plantings: PlantingDto[];
 };
 
 export type PlantingAttributeEditFormProps = EditPlantingAttributesProps & {
   addDateDefaultValue: string;
   removeDateDefaultValue: string;
-  multiplePlantings?: boolean;
+  widthDefaultValue: number | undefined;
+  heightDefaultValue: number | undefined;
+  planting: PlantingDto | null;
+  areaOfPlantings: boolean;
 };
 
 export function SinglePlantingAttributeForm({
   planting,
-  onAddDateChange,
-  onRemoveDateChange,
-  onDeleteClick,
-  isReadOnlyMode,
-}: EditSinglePlantingProps) {
+  ...props
+}: SinglePlantingAttributeFormProps) {
   const { plantId } = planting;
   const {
     data: plant,
@@ -77,10 +79,11 @@ export function SinglePlantingAttributeForm({
       <PlantingAttributeEditForm
         addDateDefaultValue={planting.addDate ?? ''}
         removeDateDefaultValue={planting.removeDate ?? ''}
-        onAddDateChange={onAddDateChange}
-        onRemoveDateChange={onRemoveDateChange}
-        onDeleteClick={onDeleteClick}
-        isReadOnlyMode={isReadOnlyMode}
+        widthDefaultValue={planting.sizeX}
+        heightDefaultValue={planting.sizeY}
+        areaOfPlantings={planting.isArea}
+        planting={planting}
+        {...props}
       />
     </div>
   );
@@ -88,12 +91,11 @@ export function SinglePlantingAttributeForm({
 
 export function MultiplePlantingsAttributeForm({
   plantings,
-  onAddDateChange,
-  onRemoveDateChange,
-  onDeleteClick,
-  isReadOnlyMode,
-}: EditMultiplePlantingsProps) {
+  ...props
+}: MultiplePlantingsAttributeFormProps) {
   const { t } = useTranslation(['plantings']);
+
+  const areAllPlantingsAreas = plantings.every((planting) => planting.isArea);
 
   const getCommonAddDate = () => {
     const comparisonDate = plantings[0].addDate;
@@ -107,6 +109,18 @@ export function MultiplePlantingsAttributeForm({
     return existsCommonDate ? comparisonDate : '';
   };
 
+  const getCommonWidth = () => {
+    const comparisonWidth = plantings[0].sizeX;
+    const existsCommonWidth = plantings.every((planting) => planting.sizeY === comparisonWidth);
+    return existsCommonWidth ? comparisonWidth : undefined;
+  };
+
+  const getCommonHeight = () => {
+    const comparisonHeight = plantings[0].sizeX;
+    const existsCommonHeight = plantings.every((planting) => planting.sizeY === comparisonHeight);
+    return existsCommonHeight ? comparisonHeight : undefined;
+  };
+
   return (
     <div className="flex flex-col gap-2 p-2">
       <h2>{t('plantings:heading_multiple_plantings')}</h2>
@@ -114,103 +128,129 @@ export function MultiplePlantingsAttributeForm({
       <PlantingAttributeEditForm
         addDateDefaultValue={getCommonAddDate() ?? ''}
         removeDateDefaultValue={getCommonRemoveDate() ?? ''}
-        onAddDateChange={onAddDateChange}
-        onRemoveDateChange={onRemoveDateChange}
-        onDeleteClick={onDeleteClick}
-        isReadOnlyMode={isReadOnlyMode}
-        multiplePlantings={true}
+        widthDefaultValue={getCommonWidth()}
+        heightDefaultValue={getCommonHeight()}
+        planting={null}
+        areaOfPlantings={areAllPlantingsAreas}
+        {...props}
       />
     </div>
   );
 }
 
 function PlantingAttributeEditForm({
+  widthDefaultValue,
+  heightDefaultValue,
   addDateDefaultValue,
   removeDateDefaultValue,
+  onWidthChange,
+  onHeightChange,
   onAddDateChange,
   onRemoveDateChange,
   onDeleteClick,
   isReadOnlyMode,
-  multiplePlantings = false,
+  areaOfPlantings,
+  planting,
 }: PlantingAttributeEditFormProps) {
   const { t } = useTranslation(['plantings']);
+  const multiplePlantings = planting === null;
 
-  const { register, handleSubmit, watch, formState } = useForm<PlantingDateAttribute>({
+  const formInfo = useForm<PlantingFormData>({
     // The 'empty' value for the native date input is an empty string, not null | undefined
     defaultValues: {
+      sizeX: widthDefaultValue,
+      sizeY: heightDefaultValue,
       addDate: addDateDefaultValue,
       removeDate: removeDateDefaultValue,
     },
     resolver: zodResolver(PlantingAttributeEditFormSchema),
   });
 
-  const addDateSubmitState = useDebouncedSubmit<PlantingDateAttribute>(
-    watch('addDate'),
-    handleSubmit,
-    onAddDateChange,
-  );
+  const { data: plant } = useFindPlantById({
+    plantId: planting?.plantId ?? 0,
+    enabled: !multiplePlantings,
+  });
 
-  const removeDateSubmitState = useDebouncedSubmit<PlantingDateAttribute>(
-    watch('removeDate'),
-    handleSubmit,
-    onRemoveDateChange,
-  );
+  const individualPlantSize = plant ? getPlantWidth(plant) : undefined;
+  const plantCountInfo =
+    planting && individualPlantSize
+      ? calculatePlantCount(individualPlantSize, planting.sizeX, planting.sizeY)
+      : null;
 
   return (
-    <>
+    <FormProvider {...formInfo}>
       {/**
        * See https://github.com/orgs/react-hook-form/discussions/7111
        * @ts-expect-error this error path was added by zod refine(). hook form is unaware, which is a shortcoming.*/}
-      {formState.errors.dateRelation && (
+      {formInfo.formState.errors.dateRelation && (
         <div className="text-sm text-red-400">{t('plantings:error_invalid_add_remove_date')}</div>
       )}
-      <div className="flex gap-2">
-        <SimpleFormInput
-          type="date"
-          id="addDate"
-          disabled={isReadOnlyMode}
-          aria-invalid={addDateSubmitState === 'error'}
-          labelContent={t('plantings:add_date')}
-          register={register}
-          className="w-36"
-        />
-        {addDateSubmitState === 'loading' && (
-          <CircleDottedIcon className="mb-3 mt-auto h-5 w-5 animate-spin text-secondary-400" />
-        )}
-        {addDateSubmitState === 'idle' && (
-          <CheckIcon
-            className="mb-3 mt-auto h-5 w-5 text-primary-400"
-            data-testid="planting-attribute-edit-form__add-date-idle"
+
+      {areaOfPlantings && (
+        <div className="py-2">
+          <h3>Area of Plantings</h3>
+
+          {plantCountInfo && (
+            <div className="grid grid-cols-[2fr_1fr] grid-rows-3 gap-y-2 py-2 text-sm">
+              <span className="font-medium">
+                {t('plantings:area_of_plantings.individual_plant_size')}
+              </span>
+              <span className="text-neutral-500">{individualPlantSize}</span>
+              <span className="font-medium">{t('plantings:area_of_plantings.plants_per_row')}</span>
+              <span className="text-neutral-500">{plantCountInfo.perRow}</span>
+              <span className="font-medium">{t('plantings:area_of_plantings.number_of_rows')}</span>
+              <span className="text-neutral-500">{plantCountInfo.perColumn}</span>
+              <span className="font-medium">{t('plantings:area_of_plantings.total_plants')}</span>
+              <span className="text-neutral-500">{plantCountInfo.total}</span>
+            </div>
+          )}
+          <DebouncedSimpleFormInput
+            onValid={onWidthChange}
+            type="number"
+            id="sizeX"
+            data-testid="planting-attribute-edit-form__size-x"
+            disabled={isReadOnlyMode}
+            labelContent={t('plantings:width')}
+            className="w-36"
           />
-        )}
-      </div>
+
+          <DebouncedSimpleFormInput
+            onValid={onHeightChange}
+            type="number"
+            id="sizeY"
+            data-testid="planting-attribute-edit-form__size-y"
+            disabled={isReadOnlyMode}
+            labelContent={t('plantings:height')}
+            className="w-36"
+          />
+          <hr className="my-4 border-neutral-700" />
+        </div>
+      )}
+
+      <DebouncedSimpleFormInput
+        onValid={onAddDateChange}
+        type="date"
+        id="addDate"
+        data-testid="planting-attribute-edit-form__add-date"
+        disabled={isReadOnlyMode}
+        labelContent={t('plantings:add_date')}
+        className="w-36"
+      />
       <p className="pb-4 text-[0.8rem] text-neutral-400">
         {multiplePlantings
           ? t('plantings:add_date_description_multiple_plantings')
           : t('plantings:add_date_description')}
       </p>
 
-      <div className="flex gap-2">
-        <SimpleFormInput
-          type="date"
-          id="removeDate"
-          disabled={isReadOnlyMode}
-          aria-invalid={removeDateSubmitState === 'error'}
-          labelContent={t('plantings:remove_date')}
-          register={register}
-          className="w-36"
-        />
-        {removeDateSubmitState === 'loading' && (
-          <CircleDottedIcon className="mb-3 mt-auto h-5 w-5 animate-spin text-secondary-400" />
-        )}
-        {removeDateSubmitState === 'idle' && (
-          <CheckIcon
-            className="mb-3 mt-auto h-5 w-5 text-primary-400"
-            data-testid="planting-attribute-edit-form__removed-on-idle"
-          />
-        )}
-      </div>
-
+      <DebouncedSimpleFormInput
+        onValid={onRemoveDateChange}
+        type="date"
+        id="removeDate"
+        data-testid="planting-attribute-edit-form__removed-date"
+        disabled={isReadOnlyMode}
+        labelContent={t('plantings:remove_date')}
+        className="w-36"
+      />
       <p className="text-[0.8rem] text-neutral-400">
         {multiplePlantings
           ? t('plantings:remove_date_description_multiple_plantings')
@@ -228,6 +268,6 @@ function PlantingAttributeEditForm({
       >
         {multiplePlantings ? t('plantings:delete_multiple_plantings') : t('plantings:delete')}
       </SimpleButton>
-    </>
+    </FormProvider>
   );
 }